Update tests & add new ones & config for tests & shell scripts
