+++ bin/tests/003_HOSTED_PAGE/0_admin/0300-HOSTED_MOTO_ADMIN.js
