+++ bin/tests/003_HOSTED_PAGE/0_admin/0301-HOSTED_NORMAL_ADMIN.js
