+++ bin/tests/003_HOSTED_PAGE/1_frontend/0300-HOSTED_IFRAME_ON_FRONTEND.js
