+++ bin/tests/011_BELFIUS/1_frontend/1100-BELFIUS_DIRECT_NET_FRONTEND.js
