+++ bin/tests/012_GIROPAY/1_frontend/1200-GIROPAY_FRONTEND.js
